+++ certifikit-cli/src/jvmMain/kotlin/app/cash/certifikit/cli/certs.kt
@@ -18,7 +18,7 @@ package app.cash.certifikit.cli
 import app.cash.certifikit.Certificate
 import app.cash.certifikit.CertificateAdapters
 import app.cash.certifikit.cli.errors.UsageException
-import app.cash.certifikit.text.certificatePem
+import app.cash.certifikit.certificatePem
 import java.security.cert.X509Certificate
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.withContext
