+++ certifikit-cli/src/jvmMain/kotlin/app/cash/certifikit/cli/host.kt
@@ -21,7 +21,7 @@ import app.cash.certifikit.cli.errors.UsageException
 import app.cash.certifikit.cli.oscp.OcspResponse
 import app.cash.certifikit.cli.oscp.ocsp
 import app.cash.certifikit.cli.oscp.toCertificate
-import app.cash.certifikit.text.certificatePem
+import app.cash.certifikit.certificatePem
 import java.io.IOException
 import java.net.InetAddress
 import kotlinx.coroutines.Deferred
