+++ certifikit/src/jvmMain/kotlin/app/cash/certifikit/pem.kt
@@ -16,11 +16,8 @@
 
 @file:JvmName("Pem")
 
-package app.cash.certifikit.text
+package app.cash.certifikit
 
-import app.cash.certifikit.AlgorithmIdentifier
-import app.cash.certifikit.CertificateAdapters
-import app.cash.certifikit.PrivateKeyInfo
 import java.security.GeneralSecurityException
 import java.security.KeyFactory
 import java.security.KeyPair
@@ -53,7 +50,7 @@ import okio.ByteString.Companion.toByteString
  * -----END CERTIFICATE-----
  * ```
  */
-actual fun String.decodeCertificatePem(): X509Certificate {
+fun String.decodeCertificatePem(): X509Certificate {
   try {
     val certificateFactory = CertificateFactory.getInstance("X.509")
     val certificates = certificateFactory
@@ -75,7 +72,7 @@ actual fun String.decodeCertificatePem(): X509Certificate {
  *
  * [rfc_7468]: https://tools.ietf.org/html/rfc7468
  */
-actual fun X509Certificate.certificatePem(): String {
+fun X509Certificate.certificatePem(): String {
   return buildString {
     append("-----BEGIN CERTIFICATE-----\n")
     encodeBase64Lines(encoded.toByteString())
@@ -89,7 +86,7 @@ actual fun X509Certificate.certificatePem(): String {
  * [rfc_5208]: https://tools.ietf.org/html/rfc5208
  * [rfc_7468]: https://tools.ietf.org/html/rfc7468
  */
-actual fun PrivateKey.privateKeyPkcs8Pem(): String {
+fun PrivateKey.privateKeyPkcs8Pem(): String {
   return buildString {
     append("-----BEGIN PRIVATE KEY-----\n")
     encodeBase64Lines(encoded.toByteString())
@@ -103,7 +100,7 @@ actual fun PrivateKey.privateKeyPkcs8Pem(): String {
  * [rfc_8017]: https://tools.ietf.org/html/rfc8017
  * [rfc_7468]: https://tools.ietf.org/html/rfc7468
  */
-actual fun PrivateKey.privateKeyPkcs1Pem(): String {
+fun PrivateKey.privateKeyPkcs1Pem(): String {
   check(this is RSAPrivateKey) { "PKCS1 only supports RSA keys" }
   return buildString {
     append("-----BEGIN RSA PRIVATE KEY-----\n")
@@ -112,7 +109,7 @@ actual fun PrivateKey.privateKeyPkcs1Pem(): String {
   }
 }
 
-internal actual fun PrivateKey.pkcs1Bytes(): ByteString {
+fun PrivateKey.pkcs1Bytes(): ByteString {
   val decoded = CertificateAdapters.privateKeyInfo.fromDer(this.encoded.toByteString())
   return decoded.privateKey
 }
@@ -149,7 +146,7 @@ internal actual fun PrivateKey.pkcs1Bytes(): ByteString {
  * [rfc_7468]: https://tools.ietf.org/html/rfc7468
  * [rfc_5208]: https://tools.ietf.org/html/rfc5208
  */
-actual fun decode(certificateAndPrivateKeyPem: String): Pair<KeyPair, X509Certificate> {
+fun decode(certificateAndPrivateKeyPem: String): Pair<KeyPair, X509Certificate> {
   var certificatePem: String? = null
   var pkcs8Base64: String? = null
   var pkcs1Base64: String? = null
@@ -204,7 +201,7 @@ actual fun decode(certificateAndPrivateKeyPem: String): Pair<KeyPair, X509Certif
   }
 }
 
-internal actual fun decodePkcs8(data: ByteString, keyAlgorithm: String): PrivateKey {
+internal fun decodePkcs8(data: ByteString, keyAlgorithm: String): PrivateKey {
   try {
     val keyFactory = KeyFactory.getInstance(keyAlgorithm)
     val x = CertificateAdapters.privateKeyInfo.fromDer(data)
@@ -214,12 +211,21 @@ internal actual fun decodePkcs8(data: ByteString, keyAlgorithm: String): Private
   }
 }
 
-internal actual fun decodePkcs1(data: ByteString): PrivateKey {
+internal fun decodePkcs1(data: ByteString): PrivateKey {
   try {
     val privateKeyInfo = PrivateKeyInfo(0L, AlgorithmIdentifier("1.2.840.113549.1.1.1", null), data)
     val pkcs8data = CertificateAdapters.privateKeyInfo.toDer(privateKeyInfo)
     return decodePkcs8(pkcs8data, "RSA")
   } catch (e: GeneralSecurityException) {
     throw IllegalArgumentException("failed to decode private key", e)
+  }
+}
+
+internal val PEM_REGEX = Regex("""-----BEGIN ([!-,.-~ ]*)-----([^-]*)-----END \1-----""")
+
+internal fun StringBuilder.encodeBase64Lines(data: ByteString) {
+  val base64 = data.base64()
+  for (i in base64.indices step 64) {
+    append(base64, i, minOf(i + 64, base64.length)).append('\n')
   }
 }
